feat(backend): implement background worker queue for processing AI feedback (fixes #317)
